rearrange code

+++ fast-stoi/src/constants.rs
@@ -1,6 +1,6 @@
 //! STOI constants
 
-pub const DYNAMIC_RANGE: f64 = 40.0;
+pub const DYNAMIC_RANGE: f32 = 40.0;
 
 // Audio frame length
 pub const FRAME_LENGTH: usize = 256; // For stft and energy as well
@@ -19,4 +19,4 @@ pub const NUM_BANDS: usize = 15; // Amount of 13 octave band
 
 pub const SEGMENT_LENGTH: usize = 30;
 
-pub const BETA: f64 = -15.0; // Lower SDR bound
+pub const BETA: f32 = -15.0; // Lower SDR bound
