+++ fast-stoi/src/stft.rs
@@ -10,22 +10,22 @@ use realfft::{RealFftPlanner, RealToComplex};
 use crate::constants::{FFT_BINS, FFT_LENGTH};
 
 lazy_static! {
-    static ref R2C: Arc<dyn RealToComplex<f64>> =
-        RealFftPlanner::<f64>::new().plan_fft_forward(FFT_LENGTH);
+    static ref R2C: Arc<dyn RealToComplex<f32>> =
+        RealFftPlanner::<f32>::new().plan_fft_forward(FFT_LENGTH);
 }
 
 /// Compute the RFFT of each valid frame as indicated by the mask.
 /// Input frames have shape (frame_length, n_frames).
 /// Returns a real valued squared magnitude spectrogram
 /// of shape (FFT_BINS, frames).
-pub fn compute_frame_rffts(frames: MatRef<f64>, mask: ColRef<bool>, count: usize) -> Mat<f64> {
+pub fn compute_frame_rffts(frames: MatRef<f32>, mask: ColRef<bool>, count: usize) -> Mat<f32> {
     // Create buffers
     let mut scratch_buffer = R2C.make_scratch_vec();
     let mut input_buffer = R2C.make_input_vec();
     let mut output_buffer = R2C.make_output_vec();
 
     // Create output array as column-major for faster writes
-    let mut spectrogram = Mat::<f64>::zeros(FFT_BINS, count);
+    let mut spectrogram = Mat::<f32>::zeros(FFT_BINS, count);
     let mut index = 0; // destination row index (skips invalid frames)
 
     // Iterate over valid frames and compute their RFFT
